refactor(cli): resource add part
